scanner: Split ScanPathTest into separate tests per scanner
This way it is easier to run only tests for a specific scanner.
