+++ scanner/src/funTest/kotlin/scanners/AbstractScannerTest.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) 2017-2019 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.ort.scanner.scanners
+
+import com.here.ort.model.config.ScannerConfiguration
+import com.here.ort.utils.safeDeleteRecursively
+
+import io.kotlintest.Spec
+import io.kotlintest.specs.StringSpec
+
+import java.io.File
+
+abstract class AbstractScannerTest : StringSpec() {
+    protected val config = ScannerConfiguration()
+    protected val licenseFilePath = File("../LICENSE")
+    protected lateinit var outputDir: File
+
+    override fun beforeSpec(spec: Spec) {
+        super.beforeSpec(spec)
+        outputDir = createTempDir()
+    }
+
+    override fun afterSpec(spec: Spec) {
+        outputDir.safeDeleteRecursively(force = true)
+        super.afterSpec(spec)
+    }
+}
