+++ scanner/src/funTest/kotlin/scanners/ScanCodeScannerTest.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2017-2019 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
+package com.here.ort.scanner.scanners
+
+import com.here.ort.utils.test.ExpensiveTag
+import com.here.ort.utils.test.ScanCodeTag
+
+import io.kotlintest.shouldBe
+
+class ScanCodeScannerTest : AbstractScannerTest() {
+    init {
+        "ScanCode recognizes our own LICENSE".config(tags = setOf(ExpensiveTag, ScanCodeTag)) {
+            val scanner = ScanCode("ScanCode", config)
+            val resultsFile = outputDir.resolve("${scanner.scannerName}.${scanner.resultFileExt}")
+
+            val result = scanner.scanPath(licenseFilePath, resultsFile)
+
+            resultsFile.isFile shouldBe true
+            result.summary.fileCount shouldBe 1
+            result.summary.licenses shouldBe sortedSetOf("Apache-2.0")
+        }
+    }
+}
